fix AppVeyor

+++ test/runtests.jl
@@ -2,7 +2,7 @@ using RequiredKeywords
 using Base.Test
 
 # Assignment form
-@required_keywords f(a; x, y::Int64, z::Int64=2) = a * x * y * z
+@required_keywords f(a; x, y::Int, z::Int=2) = a * x * y * z
 
 @test f(2, x=2, y=2,z=2) == 16
 @test f(2, x=2, y=2) == 16
@@ -10,7 +10,7 @@ using Base.Test
 @test_throws TypeError f(2, x=2, y=2.0)
 
 # Standard form
-@required_keywords function g(a; x, y::Int64, z::Int64=2)
+@required_keywords function g(a; x, y::Int, z::Int=2)
     a * x * y * z
 end
